bot/requirements.txt
python-telegram-bot==21.5
anthropic==0.34.2
supabase==2.28.3
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
